requirements.txt
smbus2
pyyaml
